+++ src/main/java/com/eaze/model/Asset.java
@@ -3,6 +3,8 @@ package com.eaze.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class Asset {
@@ -11,8 +13,8 @@ public class Asset {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private double quantity;
-    private double buyPrice;
+    private BigDecimal quantity;
+    private BigDecimal buyPrice;
 
     @ManyToOne
     private Coin coin;
