+++ src/main/java/com/eaze/service/domain/WalletService.java
@@ -4,15 +4,17 @@ import com.eaze.model.Order;
 import com.eaze.model.User;
 import com.eaze.model.Wallet;
 
+import java.math.BigDecimal;
+
 public interface WalletService {
 
     Wallet getUserWallet(User user);
 
-    Wallet addBalance(Wallet wallet, Long money);
+    Wallet addBalance(Wallet wallet, BigDecimal money);
 
     Wallet findWalletById(Long id) throws Exception;
 
-    Wallet walletToWalletTransfer(User sender, Wallet receiverWallet, Long amount) throws Exception;
+    Wallet walletToWalletTransfer(User sender, Wallet receiverWallet, BigDecimal amount) throws Exception;
 
     Wallet payOrderPayment(Order order, User user) throws Exception;
